refactor: improvements to features/FeatureFlags
