+++ src/features/FeatureFlags/index.tsx
@@ -1,20 +1,10 @@
-import React from 'react'
-import useFeatureFlags, { FeatureFlagContext } from './useFeatureFlags'
+import Provider, { useFeatureFlagsContext } from './Provider'
 
 export * from './useFeatureFlags'
 
-interface FeatureFlagProviderProps {
-  children: React.ReactNode
-}
-
-export function FeatureFlagProvider({ children }: FeatureFlagProviderProps) {
-  const featureFlags = useFeatureFlags()
-
-  return <FeatureFlagContext.Provider value={featureFlags}>{children}</FeatureFlagContext.Provider>
-}
-
 const FeatureFlagsAPI = {
-  Provider: FeatureFlagProvider
+  Provider,
+  useContext: useFeatureFlagsContext
 }
 
 export default FeatureFlagsAPI
